ajuste

+++ Models/ChequeEmpresarial.cs
@@ -10,15 +10,15 @@ namespace calculadora_api.Models
         public int Id { get; set; }
         public string dataBase { get; set; }
         public string indiceDB { get; set; }
-        public float indiceDataBase { get; set; }
+        public double indiceDataBase { get; set; }
         public string indiceBA { get; set; }
-        public float indiceDataBaseAtual { get; set; }
+        public double indiceDataBaseAtual { get; set; }
         public string dataBaseAtual { get; set; }
-        public float valorDevedor { get; set; }
+        public double valorDevedor { get; set; }
         public string encargosMonetarios { get; set; }
-        public float lancamentos { get; set; }
+        public double lancamentos { get; set; }
         public string tipoLancamento { get; set; }
-        public float valorDevedorAtualizado { get; set; }
+        public double valorDevedorAtualizado { get; set; }
         public string contractRef { get; set; }
         public string ultimaAtualizacao { get; set; }
         public string infoParaCalculo { get; set; }
